requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
gnupg==2.3.1
Markdown==3.4.1
psutil==5.9.2
pytz==2022.2.1
sqlparse==0.4.2
